Move GtkMenu to menu.go
